requirement.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
faiss-cpu
openai
python-dotenv
